make sure virtual devices are configured on remote machines
The main motivation for this is analog devices that have a min/max
value that differs from 0/1 respectively. If they are not configured
on the remote machine the input values are not correctly normalized.
As the FIXME explains there is no easy way to have the config()
member function called on all base types, so it is only done explicitly
for Analog.